requirements.txt
# Common dependencies for Poultry Farm Manager
# Note: Each service has its own requirements.txt with specific dependencies

# MQTT Client
paho-mqtt==1.6.1

# InfluxDB Client (used by monitor, analyzer, executor)
influxdb-client>=1.38.0

# Requests (used by analyzer, planner, configuration)
requests==2.31.0

# Flask (used by configuration service)
Flask>=2.3.0
